Restructures class location
