+++ classes/sequelize.js
